move state & utils to lib
